Lesson 11/requirements.txt
groq
python-dotenv
requests
structlog
fastapi
uvicorn
httpx
psutil
tenacity
